pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spoon-ai-sdk" # Package name, can be changed
version = "0.3.4" # Initial version
authors = [
  { name="Your Name", email="your.email@example.com" }, # Please replace with your info
]
description = "SDK for SpoonAI tools and agents" # A brief description
readme = "README.md" # If you have a README file
# packages = ["spoon_ai"] # REMOVED: Invalid field here
requires-python = ">=3.12" # Specify supported Python version
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License", # Choose an appropriate license
    "Operating System :: OS Independent",
]
dependencies = [ # Complete dependencies from requirements.txt for out-of-the-box usage
    "aiohappyeyeballs>=2.4.4",
    "aiohttp>=3.13.2",
    "aiosignal>=1.3.2",
    "annotated-types>=0.7.0",
    "anyio>=4.8.0",
    "attrs>=25.1.0",
    "certifi>=2025.1.31",
    "charset-normalizer>=3.4.1",
    "distro>=1.9.0",
    "fastapi>=0.115.7",
    "frozenlist>=1.5.0",
    "greenlet>=3.1.1",
    "h11>=0.14.0",
    "httpcore>=1.0.7",
    "httpx>=0.28.1",
    "idna>=3.10",
    "jiter>=0.5.0",
    "jsonpointer>=3.0.0",
    "multidict>=6.1.0",
    "openai>=1.70.0",
    "orjson>=3.10.15",
    "propcache>=0.2.1",
    "pydantic>=2.10.4",
    "pydantic-core>=2.27.2",
    "python-dotenv>=1.0.1",
    "regex>=2024.11.6",
    "requests>=2.32.3",
    "requests-toolbelt>=1.0.0",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.38",
    "starlette>=0.45.3",
    "tenacity>=8.5.0",
    "tiktoken>=0.8.0",
    "tqdm>=4.67.1",
    "typing-extensions>=4.12.2",
    "urllib3>=2.3.0",
    "yarl>=1.18.3",
    "websockets==15.0.1",
    "termcolor>=3.0.1",
    "pinecone>=6.0.2",
    "google>=3.0.0",
    "google-genai>=1.11.0",
    "protobuf>=3.19.5",
    "google-api-core>=2.24.2",
    "grpcio>=1.71.0",
    "neo-mamba>=3.0.1,<3.2.0; python_version < '3.14'",
    "neo-mamba>=3.2.0; python_version == '3.14'",
    "web3==7.11.0",
    "nest-asyncio>=1.6.0",
    "anthropic>=0.42.0",
    "boto3==1.35.99",
    "botocore==1.35.99",
    "fastmcp>=2.12.0",
    "googleapis-common-protos",
    "cryptography>=41.0.0",
    "eth-utils>=2.0.0",
    "rlp>=3.0.0",
    "uvicorn[standard]>=0.32.0",
    "x402>=1.0.0",
]

[project.optional-dependencies]
memory = [
    "mem0ai>=0.0.1",
]

[project.urls]
"Homepage" = "https://github.com/XSpoonAi/spoon-core" # Project URL
"Bug Tracker" = "https://github.com/XSpoonAi/spoon-core/issues" # Project issue tracker URL

# ADDED: Setuptools configuration to specify the package
[tool.setuptools.packages.find]
where = ["."]  # Look in the project root directory
include = ["spoon_ai*"] # Include only the 'spoon_ai' package and its subpackages
exclude = ["tests*", "test*", "cookbook*", "api*", "cli*", "agents*", "models*", "chat_logs*", "migrations*", "react_logs*", "notebooks*"] # Explicitly exclude other top-level dirs
